Switch to maven, to get the benefit of the appengine maven plugin.
This necessitated getting rid of module code isolation, for bad
reasons not worth spending time on here.

common module compiles under maven
mvn test works w/ the common module (and junit5)
tests run under idea
mavenized convert module
gdrive mavenization
service is mavenized
rm bazel files :(
mavenization: make a service war
move all code into one top-level package, sigh
more cleanup
working local dev server, with appengine maven plugin
more cleanup
