--- common/javasrc/gdocmd/common/config/storage/ServicePropertiesStorage.java
@@ -1,8 +1,0 @@
-package gdocmd.common.config.storage;
-
-import java.util.Optional;
-
-public interface ServicePropertiesStorage {
-  Optional<ServiceProperties> load();
-  void save(ServiceProperties serviceProperties);
-}
