+++ src/main/java/gdocmd/common/Functions.java
