+++ src/main/java/gdocmd/convert/Functions.java
