+++ src/main/java/gdocmd/storage/ServicePropertiesStorage.java
@@ -1,0 +1,8 @@
+package gdocmd.storage;
+
+import java.util.Optional;
+
+public interface ServicePropertiesStorage {
+  Optional<ServiceProperties> load();
+  void save(ServiceProperties serviceProperties);
+}
