+++ src/main/java/gdocmd/storage/SimulatedServicePropertiesStorage.java
@@ -1,0 +1,17 @@
+package gdocmd.storage;
+
+import java.util.Optional;
+
+public class SimulatedServicePropertiesStorage implements ServicePropertiesStorage {
+  public Optional<ServiceProperties> serviceProperties = Optional.empty();
+
+  @Override
+  public Optional<ServiceProperties> load() {
+    return serviceProperties;
+  }
+
+  @Override
+  public void save(ServiceProperties serviceProperties) {
+    this.serviceProperties = Optional.of(serviceProperties);
+  }
+}
