+++ src/test/java/gdocmd/convert/TestFunctions.java
