--- tools/bazel.rc
@@ -1,1 +1,0 @@
- build --incompatible_disallow_set_constructor=false
